feat: Remove profile option
This commit removes the `profile` option from both the `DvpmOption` and `PlugOption` types, as well as the associated logic.

The profile option was used to output the time taken to source each plugin to a buffer, but this functionality has been removed.
This commit also bumps the version number to 5.7.0.