feat: bump to v7.0.0
This commit includes the following changes:

- Bumped the version to 7.0.0 in deno.json.
- Fixed a bug in plugin.ts where `p.info.clone` was not correctly initialized.
  The condition to check clone is now `p.info.enabled ? p.info.enabled : p.info.clone`
- Changed the default value of `clone` in `PlugSchema` to `false` in types.ts.

BREAKING CHANGE: The default value for the `clone` option in the plugin configuration has been changed from `true` to `false`. This means that plugins will no longer be cloned by default, and users will need to explicitly set the `clone` option to `true` if they want a plugin to be cloned.